pyproject.toml
[project]
name = "gptzip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.2.0",
    "diskcache>=5.6.3",
    "fugashi>=1.4.0",
    "hydra-core>=1.3.2",
    "joblib>=1.4.2",
    "optimum-quanto>=0.2.6",
    "pandera>=0.22.1",
    "polars>=1.21.0",
    "protobuf>=5.29.0",
    "pydantic>=2.10.4",
    "pytest>=8.3.4",
    "quanto>=0.2.0",
    "scalene>=1.5.49",
    "torch>=2.5.1",
    "torcheval>=0.0.7",
    "transformers>=4.46.3",
    "unidic-lite>=1.0.8",
]
